Soo many changes, check long description
== Dependency changes

- Added jline 4.0.2
- Updated SnowFrame to 0.2.0
- Updated FastUtil to 8.5.18
- Updated clay4j to 0.1.3

== Code changes

- Added support for the ShiningStar updater service
- Set ShiningStar updater service to default updater service
- Added game location settings
- Added logging to the process console
- Enabled text clipping
- Miscellaneous changes due to updating ImGUI
- Adjusted some UI code due to fixed bugs in clay4j
- Added ability to auto-scroll text if its too big for the box
- Added native file dialog utils
- Changed code transformer suffix from 'Template' to '_'
- Added support for multiple products
- Updated version to 1.4.0

+++ src/main/java/de/brickforceaurora/launcher/ui/clay/FontWrapper.java
@@ -38,7 +38,7 @@ public final class FontWrapper implements IFont {
     }
 
     @Override
-    public void calculateSize(final String text, final int fontSize, final float[] size) {
+    public void calculateSize(final String text, final float fontSize, final float[] size) {
         font.calcTextSizeA(this.size, fontSize, Float.MAX_VALUE, 0.0f, text);
         size[0] = this.size.x;
         size[1] = this.size.y;
